test: split slugify into a Prisma-free module for standalone unit testing
lib/services/slug-service.ts imports @/lib/prisma at module load, which
throws without DATABASE_URL, so the pure slugify() could not be unit-tested
in isolation. Move slugify to lib/services/slug.ts (no Prisma import) and
re-export it from slug-service.ts for existing callers.
