+++ lib/services/__tests__/slug-service.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { slugify } from "../slug-service";
+import { slugify } from "../slug";
 
 // ---------------------------------------------------------------------------
 // slugify
