+++ lib/services/slug-service.ts
@@ -1,4 +1,7 @@
 import { prisma } from "@/lib/prisma";
+import { slugify } from "@/lib/services/slug";
+
+export { slugify } from "@/lib/services/slug";
 
 // Derived from `prisma.$transaction`'s own callback parameter, the same
 // pattern `case-batch-update-service.ts` uses — `Prisma.TransactionClient`
@@ -11,23 +14,6 @@ type TransactionClient = TransactionCallback extends (
 ) => Promise<unknown>
 	? T
 	: never;
-
-/**
- * Converts a name into a URL-safe slug base: lowercase, non-alphanumerics
- * collapsed to single hyphens, leading/trailing hyphens trimmed. An
- * all-punctuation/empty input falls back to "item" so a slug is always
- * produced (ADR 0003 §6 requires every published item to have one). Kept in
- * lockstep with the equivalent backfill expression in this feature's
- * migration (`prisma/migrations/20260716000000_publishing_schema_and_state_model`).
- */
-export function slugify(name: string): string {
-	const base = name
-		.toLowerCase()
-		.trim()
-		.replace(/[^a-z0-9]+/g, "-")
-		.replace(/^-+|-+$/g, "");
-	return base || "item";
-}
 
 /**
  * Generates a slug for `name` that is unique among existing CURRENT
